Remove `NavEntry` abstraction; some code cleanups
