--- src/lib/components/column-profile/NavEntry.svelte
@@ -1,77 +1,0 @@
-<script lang="ts">
-  // FIXME: we should probably rename this to AssetNavigationElement.svelte or something like that.
-  import { createEventDispatcher } from "svelte";
-  import CaretDownIcon from "$lib/components/icons/CaretDownIcon.svelte";
-  import ExpanderButton from "$lib/components/column-profile/ExpanderButton.svelte";
-
-  import { createCommandClickAction } from "$lib/util/command-click-action";
-  import { createShiftClickAction } from "$lib/util/shift-click-action";
-  import { EntityType } from "$common/data-modeler-state-service/entity-state-service/EntityStateService";
-
-  const { commandClickAction } = createCommandClickAction();
-  const { shiftClickAction } = createShiftClickAction();
-  const dispatch = createEventDispatcher();
-
-  export let entityType: EntityType;
-  export let expanded = true;
-  export let selected = false;
-  export let hovered = false;
-
-  const commandClickHandler = () => {
-    if (entityType == EntityType.Table) {
-      dispatch("query");
-    }
-  };
-
-  const clickHandler = () => {
-    dispatch("select");
-    if (
-      entityType == EntityType.Table ||
-      (entityType == EntityType.Model && selected)
-    ) {
-      dispatch("expand");
-    }
-  };
-</script>
-
-<div
-  on:mouseenter={() => {
-    hovered = true;
-  }}
-  on:mouseleave={() => {
-    hovered = false;
-  }}
-  style:height="24px"
-  style:grid-template-columns="[left-control] max-content [body] auto
-  [contextual-information] max-content"
-  class=" grid grid-flow-col gap-2 items-center hover:bg-gray-200 pl-4 pr-4 {selected
-    ? 'bg-gray-100'
-    : 'bg-transparent'}
-    "
->
-  <ExpanderButton rotated={expanded} on:click={() => dispatch("expand")}>
-    <CaretDownIcon size="14px" />
-  </ExpanderButton>
-  <button
-    use:commandClickAction
-    on:command-click={commandClickHandler}
-    use:shiftClickAction
-    on:shift-click
-    on:click={clickHandler}
-    on:focus={() => (hovered = true)}
-    on:blur={() => (hovered = false)}
-    style:grid-column="body"
-    style:grid-template-columns="[icon] max-content [text] 1fr"
-    class="w-full justify-start text-left grid items-center p-0"
-  >
-    <div
-      style:grid-column="text"
-      class="w-full justify-self-auto text-ellipsis overflow-hidden whitespace-nowrap"
-    >
-      <slot />
-    </div>
-  </button>
-  <div style:grid-column="contextual-information" class="justify-self-end">
-    <slot name="contextual-information" />
-  </div>
-</div>
